@@ -1,18 +1,59 @@
-(function() {
+// = require app/components/app-settings/app-settings.module
+// = require app/components/auth/auth.module
+// = require app/components/autocomplete/autocomplete.module
+// = require app/components/dashboard/dashboard.module
+// = require app/components/filter/filter.module
+// = require app/components/indeterminate-checkbox/indeterminate-checkbox.module
+// = require app/components/marketplace/marketplace.module
+// = require app/components/member/member.module
+// = require app/components/notifications/notifications.module
+// = require app/components/pagination/pagination.module
+// = require app/components/price-info/price-info.module
+// = require app/components/product/product.module
+// = require app/components/product-category/product-category.module
+// = require app/components/project/project.module
+// = require app/components/project-question/project-question.module
+// = require app/components/project-request/project-request.module
+// = require app/components/provider/provider.module
+// = require app/components/reload-button/reload-button.module
+// = require app/components/service/service.module
+// = require app/components/service-detail/service-detail.module
+// = require app/components/service-order/service-order.module
+// = require app/components/service-request/service-request.module
+// = require app/components/settings/settings.module
+// = require app/components/status-icon/status-icon.module
+// = require app/components/tabs/tabs.module
+// = require app/components/user/user.module
+
+(function () {
   'use strict';
 
-  angular.module('app.components', [
-    'app.core',
-    'app.services',
-
-    // Third part modules
-    'smart-table',
-    'color.picker',
-    'formly',
-    'ui.sortable',
-    'ui.bootstrap',
-    'ngDraggable',
-    'highcharts-ng',
-    'hc.marked'
+  angular.module('components', [
+    'components.app-settings',
+    'components.auth',
+    'components.autocomplete',
+    'components.dashboard',
+    'components.filter',
+    'components.indeterminate-checkbox',
+    'components.marketplace',
+    'components.member',
+    'components.notifications',
+    'components.pagination',
+    'components.price-info',
+    'components.product',
+    'components.product-category',
+    'components.project',
+    'components.project-question',
+    'components.project-request',
+    'components.provider',
+    'components.reload-button',
+    'components.service',
+    'components.service-detail',
+    'components.service-order',
+    'components.service-request',
+    'components.settings',
+    'components.status-icon',
+    'components.tabs',
+    'components.user'
   ]);
 })();
