@@ -1,29 +1,6 @@
-// This is a manifest file that'll be compiled into application.js, which will include all the files
-// listed below.
-//
-// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
-// or vendor/assets/javascripts of plugins, if any, can be referenced here using a relative path.
-//
-// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file.
-//
-// Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
-// about supported directives.
-//= require jellyfish
-//= require app/core/core.module
-//= require app/states/states.module
-//= require app/components/components.module
+// = require dohrnii
+// = require app/root.module
+// = require app/common/common.module
+// = require app/components/components.module
 
-//= require app/blocks/bind-attrs/bind-attrs.module
-//= require app/blocks/bootstrap/bootstrap.module
-//= require app/blocks/directive-options/directive-options.module
-//= require app/blocks/exception/exception.module
-//= require app/blocks/logger/logger.module
-//= require app/blocks/multi-transclude/multi-transclude.module
-//= require app/blocks/pub-sub/pub-sub.module
-//= require app/blocks/recursion/recursion.module
-//= require app/blocks/router/router.module
-//= require app/blocks/state-override/state-override.module
-//= require app/resources/resources.module
-
-//= require_tree ./app
+// = require_tree ./app
