@@ -1,18 +1,10 @@
-# == Schema Information
-#
-# Table name: product_categories
-#
-#  id              :integer          not null, primary key
-#  name            :string
-#  description     :string
-#  img             :string
-#  cached_tag_list :string
-#  deleted_at      :datetime
-#  created_at      :datetime
-#  updated_at      :datetime
-#
-
 class ProductCategorySerializer < ApplicationSerializer
-  attributes :id, :name, :description, :img, :created_at, :updated_at, :deleted_at
-  attribute :tag_list, key: :tags
+  type :product_categories
+
+  attributes :name, :description, :tag_list
+  attributes :created_at, :updated_at
+
+  def tag_list
+    object.cached_tag_list.split /,\s?/
+  end
 end
