@@ -1,33 +1,25 @@
-# == Schema Information
-#
-# Table name: services
-#
-#  id         :integer          not null, primary key
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#  type       :string           not null
-#  uuid       :string           not null
-#  name       :string           not null
-#  health     :integer          default(0), not null
-#  status     :integer
-#  status_msg :string
-#
-# Indexes
-#
-#  index_services_on_type  (type)
-#  index_services_on_uuid  (uuid)
-#
-
 class ServiceSerializer < ApplicationSerializer
-  attributes :id, :type, :name, :uuid, :status, :health, :status_msg, :created_at, :updated_at
+  attributes :user_id, :provider_id, :product_id, :project_id, :service_request_id, :service_order_id
+  attributes :name, :state, :billable, :hourly_price, :monthly_price
+  attributes :last_changed_at, :last_checked_at
+  attributes :created_at, :updated_at
 
-  # Relationships
-  has_one :order, serializer: OrderSerializer
-  has_one :project, serializer: ProjectSerializer
-  has_one :product, serializer: ProductSerializer
-  has_one :product_type, serializer: ProductTypeSerializer
-  has_one :provider, serializer: ProviderSerializer
+  field :monitor_frequency
 
-  has_many :logs, serializer: LogSerializer
-  has_many :service_outputs
+  field :status_message
+  field :type_name, as: :type
+  field :settings, :details, :actions
+
+  filter :with_states
+
+  has_one :user
+  has_one :provider
+  has_one :product
+  has_one :project
+  has_one :service_request
+  has_one :service_order
+
+  def actions
+    object.actions.split ','
+  end
 end
